requirements.txt
fastapi[all]==0.95.1
uvicorn==0.21.1
beanie==1.18.0
python-dotenv==1.0.0
mypy==1.2.0
pre-commit==3.2.2
